[Aptos Data Client] Split out the client.
