+++ state-sync/state-sync-v2/state-sync-driver/src/driver_factory.rs
@@ -14,7 +14,7 @@ use crate::{
 };
 use aptos_config::config::NodeConfig;
 use aptos_consensus_notifications::ConsensusNotificationListener;
-use aptos_data_client::aptosnet::AptosNetDataClient;
+use aptos_data_client::aptosnet::client::AptosNetDataClient;
 use aptos_data_streaming_service::streaming_client::StreamingServiceClient;
 use aptos_event_notifications::{EventNotificationSender, EventSubscriptionService};
 use aptos_executor_types::ChunkExecutorTrait;
